backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
boto3==1.36.26
django-storages==1.14.5
whitenoise==6.9.0
Pillow==11.1.0
